// FRONTEND/lib/types/database.ts
export type ProgrammingLanguage =
  | "CSHARP"
  | "NODE_JS"
  | "JAVA"
  | "GO"
  | "PYTHON"
  | "TYPESCRIPT"
  | "JAVASCRIPT"
  | "PHP"
  | "RUST"
  | "KOTLIN"
  | "SWIFT"
  | "OTHER";

export type PlatformType = "WEB" | "DESKTOP" | "MOBILE";

export type ProjectStatus =
  | "OPEN"
  | "NEGOTIATING"
  | "IN_DEVELOPMENT"
  | "COMPLETED"
  | "CANCELLED";
// ...
export interface ProjectCreate {
  clientId: string;
  title: string;
  description: string;
  category: string;
  sub_category: string;
  primaryLanguage?: ProgrammingLanguage;
  platforms: PlatformType[];
  status?: ProjectStatus;
  programmerId?: string | null;
}